+++ example/lib/screens/simple_screen.dart
@@ -46,6 +46,8 @@ class _SimpleScreenState extends State<SimpleScreen> {
                 onUnityMessage: onUnityMessage,
                 onUnitySceneLoaded: onUnitySceneLoaded,
                 webUrl: 'http://localhost:6080/',
+                useAndroidViewSurface: true,
+                borderRadius: BorderRadius.all(Radius.circular(70)),
               ),
               Positioned(
                 bottom: 20,