+++ lib/src/io/device_method.dart
@@ -169,15 +169,18 @@ class MethodChannelUnityWidget extends UnityWidgetPlatform {
     }
 
     if (defaultTargetPlatform == TargetPlatform.android) {
-      if (!useAndroidViewSurface) {
-        return AndroidView(
-          viewType: _viewType,
-          onPlatformViewCreated: onPlatformViewCreated,
-          gestureRecognizers: gestureRecognizers,
-          creationParams: creationParams,
-          creationParamsCodec: const StandardMessageCodec(),
-        );
-      }
+      // commented for 3.0.0 changes in platform view rendition
+      // if (!useAndroidViewSurface) {
+      //   return AndroidView(
+      //     viewType: _viewType,
+      //     onPlatformViewCreated: onPlatformViewCreated,
+      //     gestureRecognizers: gestureRecognizers,
+      //     creationParams: creationParams,
+      //     creationParamsCodec: const StandardMessageCodec(),
+      //     hitTestBehavior: PlatformViewHitTestBehavior.opaque,
+      //     layoutDirection: TextDirection.ltr,
+      //   );
+      // }
 
       return PlatformViewLink(
         viewType: _viewType,
@@ -193,8 +196,7 @@ class MethodChannelUnityWidget extends UnityWidgetPlatform {
           );
         },
         onCreatePlatformView: (PlatformViewCreationParams params) {
-          final SurfaceAndroidViewController controller =
-              PlatformViewsService.initSurfaceAndroidView(
+          final controller = PlatformViewsService.initExpensiveAndroidView(
             id: params.id,
             viewType: _viewType,
             layoutDirection: TextDirection.ltr,
@@ -202,6 +204,7 @@ class MethodChannelUnityWidget extends UnityWidgetPlatform {
             creationParamsCodec: const StandardMessageCodec(),
             onFocus: () => params.onFocusChanged(true),
           );
+
           controller
             ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
             ..addOnPlatformViewCreatedListener(onPlatformViewCreated)