+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 2022.1.1+2
+
+* Fixed android screen offset
+* 
+## 2022.1.1+1
+
+* Fixed issue with unity crashing on ios when screen is pushed into stack
+
+## 2022.1.1
+
+* Fixed issues [#35](https://github.com/juicycleff/flutter-unity-view-widget/pull/595) by [@jamesncl](https://github.com/juicycleff/flutter-unity-view-widget/issues?q=is%3Apr+author%3Ajamesncl)
+* Force expensive surface as default for android
+
+## 2022.1.0+5
+
+* Migrated to Flutter 3.0.0
+* Issues with Android with AndroidView now exists
+
 ## 2022.1.0+4
 
 * Fixed unity screen turning white on scene load (Android)